// frontend/src/generated-api/model/friendshipStatus.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BS.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type FriendshipStatus = typeof FriendshipStatus[keyof typeof FriendshipStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FriendshipStatus = {
  Friend: 'Friend',
  IncomeRequest: 'IncomeRequest',
  OutcomeRequest: 'OutcomeRequest',
  None: 'None',
} as const;